Added html wireframe files
This HTML/Java code is based off of a portion of another project I did for a private party. This portion of the code is now obsolete in relation to that private party's wishes. It has been scrubbed of information and themes that could be used to tell what the original product was. The only thing it has in common with the original product is my own style of creating menus, which will also be changed as we progress in this project.